Title: 20대의 평균나이, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 241003/20대의 평균나이/average-age-of-20.py
@@ -5,5 +5,5 @@ while True:
         a.append(age)
     elif age >= 30:
         break
-res = float(round(sum(a)/len(a), 2))
+res = round(sum(a)/len(a), 2)
 print("%.2f" %res)
